requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.0
SQLAlchemy==2.0.23
apscheduler==3.10.4
python-dotenv==1.0.0
requests
Flask